requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Google Gemini AI
google-generativeai==0.3.2

# Sentence Transformers untuk Semantic Search
sentence-transformers==2.2.2

# PyTorch untuk operasi tensor (dibutuhkan oleh sentence-transformers)
# Gunakan versi terbaru yang tersedia di environment Anda.
torch==2.9.1

# Dependencies tambahan untuk sentence-transformers
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2

# Utilities
requests==2.31.0


# Evaluasi teks
rouge-score==0.1.2

# Excel / data tabular
pandas==2.1.4
openpyxl==3.1.2



